+++ lib/modules/state/state_service.dart
@@ -1,9 +1,9 @@
 ﻿// Module: state
-// Updated at: 2026-06-26 09:28:35
-// Commit: feat(state): add async state stream provider
+// Updated at: 2026-06-26 10:46:32
+// Commit: feat(state): integrate custom telemetry logger in state
 
 class stateComponentState {
-  final String commitRef = '2345';
-  final String lastMessage = 'feat(state): add async state stream provider';
-  final DateTime timestamp = DateTime.parse('2026-06-26T09:28:35+05:00');
+  final String commitRef = '2351';
+  final String lastMessage = 'feat(state): integrate custom telemetry logger in state';
+  final DateTime timestamp = DateTime.parse('2026-06-26T10:46:32+05:00');
 }